+++ addon/index.js
@@ -8,6 +8,16 @@ export {
   Promise
 } from './-private/concurrency-helpers';
 
+export {
+  spawn,
+  spawnChild,
+  stop,
+  current,
+  childrenSettled,
+  parallel,
+  serial
+} from './-private/scheduler';
+
 export { printSprites } from './-private/debug';
 
 export {