+++ tests/dummy/app/controllers/hero/detail.js
@@ -2,25 +2,20 @@ import Controller from '@ember/controller';
 import move, { continuePrior } from 'ember-animated/motions/move';
 import scale from 'ember-animated/motions/scale';
 import opacity from 'ember-animated/motions/opacity';
+import { parallel } from 'ember-animated';
 
 export default Controller.extend({
   transition: function * ({ receivedSprites, sentSprites, removedSprites }) {
-    receivedSprites.forEach(sprite => {
-      sprite.scale(sprite.initialBounds.width / sprite.finalBounds.width, sprite.initialBounds.height / sprite.finalBounds.height);
+
+    // received and sent sprites are flying above all the others
+    receivedSprites.concat(sentSprites).forEach(sprite => {
       sprite.applyStyles({
         'z-index': 1
       });
-      move(sprite);
-      scale(sprite);
     });
 
-    sentSprites.forEach(sprite => {
-      sprite.applyStyles({
-        'z-index': 1
-      });
-      move(sprite);
-      scale(sprite);
-    });
+    receivedSprites.forEach(parallel(move, scale));
+    sentSprites.forEach(parallel(move, scale));
 
     removedSprites.forEach(sprite => {
       sprite.endTranslatedBy(0, 0);